feat(dashboards-v2): shared scrollbar mixin + align table/editor with refactor
- styles: add a shared `custom-scrollbar` mixin (always-visible thin
  scrollbar — transparent track, l3 thumb) and use it in the chart
  Legend and the TablePanel (wrapper + antd `.ant-table-body` /
  `.ant-table-content`), so the table scrollbar matches the legend.
- table/editor: point getPanelDefinition imports at `Panels/registry`,
  the chartAppearance resolver at its new path, drop the now-required
  `panel?.`/`data?.` optional chaining, and render the shared `NoData`
  empty-state; update the renderer test accordingly.

echo "=== log ==="; git log --oneline -2; echo "=== tree ==="; git status --short
